pyproject.toml
[project]
name = "pythondojo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.13",
    "langchain-deepseek>=0.1.3",
    "langchain-google-genai>=2.1.4",
    "langgraph>=0.4.3",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
